requirements.txt
faiss>=1.7.4
pandas>=1.5.3
pdf2image>=1.16.3
Pillow>=9.5.0
pytesseract>=0.3.10
python-telegram-bot>=20.3
sentence_transformers>=2.2.2
torch>=2.0.1
transformers>=4.30
